rename SignInMainView to SignInView, remove old SignInView(with cards)
