+++ src/components/StudioEntrypoint/StudioEntrypoint.tsx
@@ -33,12 +33,6 @@ export const StudioEntrypoint: React.FC<StudioEntrypointProps> = ({ enterLocatio
   // not signed user with not created memberships and/or no extension
   if (!hasMemberships) {
     // go to /signin/join
-    return <Navigate to={absoluteRoutes.studio.signInJoin()} replace />
-  }
-
-  // not signed user with extension and with created memberships
-  if (hasMemberships && !memberSet) {
-    // go to /signin
     return <Navigate to={absoluteRoutes.studio.signIn()} replace />
   }
